fix($cluster): query tweets from all influencers
Fixes TS-38:

Here's the problem: because certain influencers (e.g. Elon Musk and Tesla) have
scores for multiple clusters.id, whenever the select distinct happens to use the
row that has a different clusters.slug than the one that we're querying for
(e.g. if the @elonmusk row uses his attention_score for DogeCoin but we're
querying for articles about Tesla, all of Elon's tweets will be omitted from the
results).

The fix: select distinct on (urls.link_id, tweets.author_id, tweets.cluster_id
so that I get all of the rows that have unique links, authors, and clusters.
Then, the remaining of the query filters for a specific cluster so no duplicate
tweets (from the same author for the same link but with a different cluster's
score) should show up.